+++ CHANGELOG.md
@@ -7,10 +7,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ## Unreleased
 
 ### Added
+- Nautilus Component: Allow the edit text based files. Using rclone rc operations/uploadfile to save.
 - Nautilus Component: Added delete, move and copy operations support.
 - Nautilus Component: Added vertical split mode support.
 
 ### Changed
+- Font Awesome icons updated to 7.2.0. Some icon changed Google Material Icons.
+- Removed Syntax Highlighting, added CodeMirror.
 - Nautilus Component: A lot of improvements and bug fixes.
 - Notification behavior improved.